floordiv review: removed is_none function; .fetchall > .fetchone; get hash of password once; minor edits

+++ server/todos/app.py
@@ -31,71 +31,79 @@ async def root():
 
 @app.post("/am/signup")
 async def am_signup(user: UserModels.UserAuth = Depends()):
-    user_handler = UserHandlers.User(mysql, user=user)
+    user_handler = UserHandlers.User(connection=mysql, user=user)
     return await user_handler.signup()
 
 
 @app.post("/am/signin")
 async def am_signin(user: UserModels.UserAuth = Depends()):
-    user_handler = UserHandlers.User(mysql, user=user)
+    user_handler = UserHandlers.User(connection=mysql, user=user)
     return await user_handler.signin()
 
 
 @app.post("/notebook/add")
 async def notebook_add(user: UserModels.UserFunc = Depends()):
-    user_handler = UserHandlers.User(mysql, user=user)
-    notebook_handler = NotebookHandlers.Notebook(mysql, user=user)
+    user_handler = UserHandlers.User(connection=mysql, user=user)
+    notebook_handler = NotebookHandlers.Notebook(connection=mysql, user=user)
 
     if not await user_handler.user_exists():
         return {"status": False}
     if not await user_handler.is_valid_token():
         return {"status": False}
+    if not bool(user.username) or not bool(user.token):
+        return {"status": False, "reason": "Один из параметров имеет <null> тип"}
     return await notebook_handler.add()
 
 
 @app.post("/notebook/update")
 async def notebook_update(data: NotebookModels.DataNotebook, user: UserModels.UserFunc = Depends()):
-    user_handler = UserHandlers.User(mysql, user=user)
-    notebook_handler = NotebookHandlers.Notebook(mysql, user=user, data=data)
+    user_handler = UserHandlers.User(connection=mysql, user=user)
+    notebook_handler = NotebookHandlers.Notebook(connection=mysql, user=user, data=data)
 
     if not await user_handler.user_exists():
         return {"status": False}
     if not await user_handler.is_valid_token():
         return {"status": False}
+    if not bool(user.username) or not bool(user.token):
+        return {"status": False, "reason": "Один из параметров имеет <null> тип"}
     return await notebook_handler.update()
 
 
 @app.post("/notebook/delete")
 async def notebook_delete(data: NotebookModels.DataNotebook, user: UserModels.UserFunc = Depends()):
-    user_handler = UserHandlers.User(mysql, user=user)
-    notebook_handler = NotebookHandlers.Notebook(mysql, user=user, data=data)
+    user_handler = UserHandlers.User(connection=mysql, user=user)
+    notebook_handler = NotebookHandlers.Notebook(connection=mysql, user=user, data=data)
             
     if not await user_handler.user_exists():
         return {"status": False}
     if not await user_handler.is_valid_token():
         return {"status": False}
+    if not bool(user.username) or not bool(user.token):
+        return {"status": False, "reason": "Один из параметров имеет <null> тип"}
     return await notebook_handler.delete()
 
 
 @app.post("/notebook/get")
 async def notebook_get(user: UserModels.UserFunc = Depends()):
-    user_handler = UserHandlers.User(mysql, user=user)
-    notebook_handler = NotebookHandlers.Notebook(mysql, user=user)
+    user_handler = UserHandlers.User(connection=mysql, user=user)
+    notebook_handler = NotebookHandlers.Notebook(connection=mysql, user=user)
 
     if not await user_handler.user_exists():
         return {"status": False, "reason": "Данного имени пользователя не существует"}
     if not await user_handler.is_valid_token():
         return {"status": False, "reason": "Токен не верный"}
+    if not bool(user.username) or not bool(user.token):
+        return {"status": False, "reason": "Один из параметров имеет <null> тип"}
     return await notebook_handler.get()
 
 
 @app.post("/am/check/token_valid")
 async def am_check_token(user: UserModels.UserFunc = Depends()):
-    user_handler = UserHandlers.User(mysql, user=user)
+    user_handler = UserHandlers.User(connection=mysql, user=user)
     return {"status": await user_handler.is_valid_token()}
 
 
 @app.post("/am/check/user_exists")
 async def am_check_user_exists(user: UserModels.User = Depends()):
-    user_handler = UserHandlers.User(mysql, user=user)
+    user_handler = UserHandlers.User(connection=mysql, user=user)
     return {"status": await user_handler.user_exists()}
